changed migration

+++ db/migrate/20181031184342_add_store_to_products.rb
@@ -2,9 +2,9 @@ class AddStoreToProducts < ActiveRecord::Migration[5.2]
   def change
     add_column :products, :storename, :string
     add_column :products, :storeurl, :string
-    add_column :products, :storefeedbackscore, :float
-    add_column :products, :storeprice, :float
+    add_column :products, :storefeedbackscore, :decimal
+    add_column :products, :storeprice, :decimal
     add_column :products, :storeepacket, :boolean
-    add_column :products, :storeepacketprice, :float
+    add_column :products, :storeepacketprice, :decimal
   end
 end
